+++ src/input_flac.h
@@ -37,7 +37,7 @@ namespace audiere {
     static FLAC__StreamDecoderReadStatus read_callback(
       const FLAC__StreamDecoder* decoder,
       FLAC__byte buffer[],
-      long unsigned int* bytes,
+      size_t *bytes,
       void* client_data);
     static FLAC__StreamDecoderSeekStatus seek_callback(
       const FLAC__StreamDecoder* decoder,
